Implement CovertSender.send( )
Previously a whole chunk of message was sent through this class, now we can call .send() to send messages separately, it still breaks the message into chunks to fit into UDP datagram.

At this point it just waits until a single packet delivered successfully, a multithreaded approach would be great to send a batch of packets and implement sliding window to allow reliability.