Extract covert bits in receiver (!)
Now the receiver checks the checksum to save the covert bit. It can restore the covert message indeed;
however, there are extra characters in the end, that should be fixed. Currently it prints:

Covert message: a covert message.ÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿÿ

Also note that only 1 bit of info can be covertly send in this manner, so the input message is made long,
so it'll be divided into chunks, and I also limited the udp payload to something small such that there'll
be more packets to send. rather than few packets with large messages.