Rename symbols
To receive the ACKs I created a socket before but in the current version it wasn't used yet.
It was a bit confusing because  the name self.sock doesn't reveal its purpose clearly
I thought it is the socket to send and receive data but apparently in order to modify
UDP checksum, scapy's send is used instead of socket module, so this variable self.sock
was dangling. Then I remembered I put it in order to also receive ACKs, so I rename it to
self.ack_sock