Fix loop bug (*)
Now the sender stops once the covert bits are all sent. I'm not sure if this
is a good behavior though because the carrier message is not fully sent to
the receiver, rather it's partially sent until the covert bits are all acknowledged.

I thought this behaviour is helpful to measure the capacity of the channel
to get informed in what time it could send the message, but that could've
been implemented differently.

Nevertheless, after this commit the logic of covert channel, in its bare bones
is done: send packets in a window, get acks / slide window, repeat until all covert bits are sent.

I put a (*) to indicate a milestone.