Fix infinite waiting of ACKs
In case receiver is not running, sneder still waits for ACK to come, that's
because the start of the window is not slided when the packet is decided
to be dropped.  Now sender can continue on other packets if ACK is not
received in expected timeout and max retransmissions is reached.