Add sleep to ACK and add note
I see time.sleep( ) was used in the given template,
looks like it's needed to avoid blocking other threads.
I realized current measured capacity has an implicit assumption
that if an ACK is not received and maximum retransmission is
achieved, it takes it as dropped packet, but it can still reach
in receiver's side, so that can be a note add in the report? Or change
the capacity measurements implementation.