Minor fix inequality self.max_trans
Apparently max_trans = 1 means only send packet once, however with the
previous inequality it actually meant retransmission of 1, i.e. 2 packets
would be sent at most. I'm not sure if this naming is good, I might change it
back to "retransmission" so that when it is set to 0, it won't resend packets.