Send covert msg with probability
Now the sender sends covert message repeadetly (so that it can be used to capture traffic as long as needed) but also with a parameter P,
where with probability P covert message is sent and with
probability 1-P only carrier is sent. I thought it might in evaluating
detection stats, to compute FP / TN / Accuracy etc.