Save parameters to .json
Untested, it should check if .json exists, if not it creates it
otherwise it loads, checks if there's an entry of the same parameters,
if not, creates a unique .csv filename and adds it to .json

This way, different parameters will create different datasets and
during the experiments we can measure e.g. accuracy of current detector
with respect to these parameters.